Fix immutable properties issue
- Immutables fields were not checked properly causing terraform
to believe that all fields were immutable
- Added roll back mechanism in case an immutable field has been updated.
In such cases, the incoming terraform data would be updated form the
tf file; but in this case the expectation is for the state file to remain
as is. Thus, the incoming data object is updated with the data from remote
making sure that the state file matches the remote state.