Add support for global security schemes
- Refactor authentication mechanism and use interfaces instead
- Auth context is generated inside prepareAuth now instead of when
creating the OperationAuthenticator, this forces the context to be used
and populated with the authentication metadata as opposed to using the
struct object which ended up being redundant in the struct.
- apiAuthenticator is now a dependency of resoureFactory instead
of creating the authenticator inside the methods. This eases testing
and also makes de struct resourceFactory design more robust in case a new
implementation of apiAuthenticator is needed in the future
- Potential secret leak in the URL is now fixed, printing the resource URL
instead of the URL from the auth context which may contain security tokens