use always terraform compliant names when reading from/writing into the state
- this change fixes the issue where properties using the extension
'x-terraform-field-name' would not get stored property into the state file
and were being stored using the property name instead. Similarly, when
saving payload into the state file the property names were being used instead
of the preferred name defined in the 'x-terraform-field-name' extension field.
This was causing errors when processing the resources as the resource schema
was correctly set up with either the property name or the preferred one but when
reading/writing to state the proper conversion was not being done and terraform
would fail.
- using getResourceDataOKExists handles Zero values too so no need to
have getResourceDataChange method anymore. This simplifies the code.
- immutable code now uses getResourceDataOKExists instead of getResourceDataValue