check errors when reading and if error is NotFound then return nil as expected
- this change is needed specially when terraform handles updates/deletions doing a
refresh first (performing a read call) in which case if an error is thrown then updates/deletes will
not work as terraform will stop the execution right away. The reading handler
will wipe out the state of a resource if its not found remotely which is
the expected behaviour

- allow 404 when checking returned status code on deletion. if 404 that means
resource has been deleted already...also even if polling is enabled, the polling
mechanism will also capture that and return successfully

- increased the timeout duration to meet service providers APIs where operations
take more than 1 min. Setting it to 10 tentatively for now.