remove symlink support
- symlinks don't resolve well in certain platforms causing errors at runtime.
This is not a reliable solution and it was just created in the first
place to optimise memory space which is not a concern right now considering
the size of the provider binary. Since this is causing more issues than benefits
 dropping it.
- make file: just install the plugin the terraform plugin folder with the provider name
- added new target examples-container that allows for examples to be
run in an encapsulated environment already set up to run terraform
- install script does not create symlinks anymore, rather it downloads and
installs the provider with the corresponding provider name.
- updated readmes, root readme, local env and examples readme

- added networks to services in the docker compose
- created a new service terraform-provider-openapi-examples that encapsulates
an environment ready to run terraform and the different examples
- removed host from swagger file so request from within the container
are router properly
- created examples terraform-provider-openapi.yaml which is used in the
container example