[BIG CHANGE] this is the alternative change using TypeMaps as expected
This change fixes the issue where the payload build from the terraform
config file (with values) was not honoring the object property types and instead
it was converting all the object property values into strings causing issues
with the API as it was not able to marshall the payload into the right
struct...

Source of the problem: The OpenAPI Terraform provider is using TypeMap
for the swagger properties that are of type 'object' and setting up
each of the properties within the object accordingly with their property
type. However, as per Terraform TypeMAP documentation typeMAps only support
as items (For a TypeMap, it must be a *Schema with a Type of TypeString)
or any other primitive type like TypeInt, TypeFloat, TypeBool for that matter.

This is a limitation as the provider is using a *Resource to be able to
define the actual schema of the object. This solution enables objects
to be translated into maps within terraform using their own type; but
the only thing to keep track of is that Terraform will save the values
of the properties using the default elem type for Maos which is strings. So
the OpenAPI provider is handling this gracefully by leveraging the metadata
for each ofthe properties in swagger and translating accordingly when
creating the payloads (honoring the property types) and also when converting
from payload to state data.

Summary of changes and things to keep in mind:

- Data from terraform configuration file to payload for object properties
are translated accordingly honoring the object properties data type (not
always being a string). This is done via handling the case reflect.String
to do the right translations when the actual property type is different.

- Data from payload to terraform configuration file goes through a conversion
process just for object types where each value will be converted into
string making terraform happy. This is done via handling specific type
conversions to string in each of the reflect primitives. This is again
only applicable to object properties (elems of TypeMAP). For other terraform
types like TypeString, TypeInt, TypeBool, TypeFloat, TypeList this is not
applicable.

- Run into a known conversion issue when un-marshalling payloads where
integers are automatically coverted to float type objects...Hence, the
need for having to cover this special case in the reflect.Float64 case
to confirm if the property type is an int and if so translate to int instead.