remove Default mapping from terraformSchema
Based on main terraform core maintainer the default value is set so users
can know at plan time what the value for a given property is if the value
is not provided. However, this goes a bit against the openapi spec since
the expected behavior for computed and optional computed is taht the API
itslef will be generating the value as opposed to having the client posting
the value.