only mark as computed when property is optional and is eiter readonly or optioanl computed
- optional computed meaning now only covers the use case where the
property is optional, not readonly, computed and the default is not set. The
reason for this decision is because for the optioanl computed where the default
value is known at plan time, terraform will push the value itself. This
provides more info to the user when planning the configuration

- the default value is only added to the terraform schema when the property is
not computed