expand schema
- with the schema now being expanded when figuring out the terraform
resources now instead of enforcing the ref to be present and then discovering
the actual schema of the pointed ref inspecting the definitions, now the
schema will be automatically embedded into the post operation body parameter.
- this leaves now the option to service providers to have a post operation
body that does not match the get operation (this should be handled appropriately
but probably out of the scope of this story). The code has been documented
as a reference for the future.