replace IsNestedObject with isPropertyWithNestedObjects method
- logic to figure out whether the property contains nested objects lies now
on the schema definition property as opposed to the field IsNestedObject representing it.
This makes sure the schema property can not be hacked with missleading info
For instance, having a schema property that has  IsNestedObject set to true
but the type being not an object and the SpecSchemaDefinition being nil.
This makes also the tests more robusts.
- removed no longer needed isPropertyWithNestedObjects
- added unit tests for TestIsPropertyWithNestedObjects