support subresource name processing from the path when building the resource name
- currently the names for subresources are build based on the path resource
names and preferred names are not supported. This should be supported
probably next so we honor the preferred names and keep things consistent. Otherwise
we may have root resources that are using the preferred names but their subresources
using the path names which will be confusing for the user.
- fixed issue where the very first version found was attached to the resource. This
was not working as expected when dealing with subresources. Now the version
is expected to be next to the resource in the path, if it is not present then
the resource is considered to not have a version
-