refactor ProviderOpenAPI and make CreateSchemaProviderWithConfiguration accept a ServiceConfiguration to facilitate unit testing
- created a unit test given a swagger file input and confirming that provider is configured properly. This
helps understand the value of the provided and provides more visibility on
how the provider works. Given a swagger file, when the ProviderOpenAPI is
created then the resulted terraform provider should be configured as expected.
- updated int test to make use of the helper method CreateSchemaProviderWithConfiguration
to configure the provider using the ServiceConfigStub with the SwaggerURL field
populated with the swagger URL where the swagger openapi doc is being served
- make ServiceConfigStub public so it can be used from int test package