add support for property description
- this commit shows how functionality not currently available can be added
so the doc renderer will be able to make use of it. In this case the attrobute
description was not supported in the SpecSchemaDefinitionProperty. So a new field
was added to the SpecSchemaDefinitionProperty struct and also the method in charge
of populating the struct from the content in the swagger file (createSchemaDefinitionProperty)
was also updated so the description field in the swagger definition property is
now also translated into the intermidiaty object SpecSchemaDefinitionProperty