create method that compares local proeprty value with remote
- this method will be in charge of deciding whether the state needs to be
updated with the property value or not. terraform saves already the input
properties and values from the tf file into the ResourceData property, so
if the values from the API matches the ones introduced by the user then the update does not
need to updated. For other properties through like computed the state needs
to be updated accordingly.