Deprecate multi-region support on resource name level
This feature was introduced in early days of the provider to support multi-region. However,
this model was not compatible with Terraform modules since the region was attached to the resource name.
Later, multiregion support was added on the provider level as expected by Terraform enabling
the provider to be used in modules and enabling users to specify the regions on the provider level.

The multi-region support on the resource name level has been still supported till now but it's time
to say goodbye to features that had some value in the past but now don't make sense and makes the code
more difficult to read and maintain.